+++ app/Http/Controllers/Admin/JobRoleController.php
@@ -88,11 +88,11 @@ class JobRoleController extends Controller
     public function destroy(string $id)
     {
         // validation
-        // $jobExist = Job::where('job_role_id', $id)->exists();
+        $jobExist = Job::where('job_role_id', $id)->exists();
 
-        // if($jobExist) {
-        //     return response(['message' => 'This item is already been used can\'t delete!'], 500);
-        // }
+        if($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
 
         try {
             JobRole::findOrFail($id)->delete();
