+++ app/Http/Controllers/Admin/JobTypeController.php
@@ -89,11 +89,11 @@ class JobTypeController extends Controller
     public function destroy(string $id)
     {
         // validation
-        // $jobExist = Job::where('job_type_id', $id)->exists();
+        $jobExist = Job::where('job_type_id', $id)->exists();
 
-        // if($jobExist) {
-        //     return response(['message' => 'This item is already been used can\'t delete!'], 500);
-        // }
+        if($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
 
         try {
             JobType::findOrFail($id)->delete();
